+++ packages/button/src/react/__specs__/index.spec.tsx
@@ -1,9 +1,8 @@
+import { CheckIcon } from '@pluralsight/ps-design-system-icon'
 import { fireEvent, render } from '@testing-library/react'
 import React from 'react'
 
-import { CheckIcon } from '@pluralsight/ps-design-system-icon'
-
-import Button from '..'
+import Button from '../index.js'
 
 describe('Button', () => {
   it('should render a button', () => {
